Update dependencies and improve macOS compatibility
- Bump `xcap` dependency version from 0.0 to 0.9 in Cargo.toml.
- Add rpath for Swift runtime in build.rs to ensure proper linking on macOS.
- Fix sample rate type in loopback.rs by casting SAMPLE_RATE to i32.

+++ src-tauri/src/services/loopback.rs
@@ -567,7 +567,7 @@ mod capture {
         // dropped. A tiny frame size keeps that overhead negligible.
         let config = SCStreamConfiguration::new()
             .with_captures_audio(true)
-            .with_sample_rate(SAMPLE_RATE)
+            .with_sample_rate(SAMPLE_RATE as i32)
             .with_channel_count(1)
             .with_width(2)
             .with_height(2);
